PYTHON_BS_PYTEST/emembership_sel/requirements.txt
attrs==23.1.0; python_version >= '3.7'
browserstack-local==1.2.7
browserstack-sdk==1.19.4
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; os_name == 'nt' and implementation_name != 'pypy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
colorama==0.4.6; sys_platform == 'win32'
dukpy==0.3.0
exceptiongroup==1.2.0; python_version >= '3.7'
faker==21.0.0
greenlet==2.0.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.3.0; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
multipledispatch==1.0.0
outcome==1.3.0.post0; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
parse==1.20.0
parse-type==0.6.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
playwright==1.40.0; python_version >= '3.8'
pluggy==1.3.0; python_version >= '3.8'
psutil==5.9.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.21
Pipenv
pyee==11.1.0
pypac==0.16.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyshadow==0.0.5; python_version >= '3.7'
pysocks==1.7.1
pytest==7.4.3; python_version >= '3.7'
pytest-bdd==7.0.1; python_version >= '3.7' and python_version < '4.0'
pytest-html==4.1.1; python_version >= '3.8'
pytest-metadata==3.0.0; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
python-dateutil==2.8.2
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
sniffio==1.3.0; python_version >= '3.7'
sortedcontainers==2.4.0
tldextract==5.1.1; python_version >= '3.7'
trio==0.23.2; python_version >= '3.7'
trio-websocket==0.11.1; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.1.0
selenium==4.16.0; python_version >= '3.7'
numpy==1.26.2
PyAutoGUI~=0.9.53
Pillow==10.1.0
imageio==2.33.1
imageio_ffmpeg==0.4.9
